+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Improved loading progress overlay repaints to avoid freezing when document became inactive
 - Changed a bit default page in model free mode
+- Fixed `children` option ignorance in `itemConfig` when `limitLines` is used for `tree` view
 - Fixed `href` option to work for `button` view
 - Fixed path generation in `signature` details
 - Added `Widget#pathToQuery(path)` method